requirements.txt
Django==4.2
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0
google-api-python-client==2.120.0
requests==2.31.0
python-dotenv==1.0.1
